requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-talisman==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
greenlet==3.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
typing-extensions==4.12.2
python-dotenv==1.0.1
cryptography==44.0.0
flask_wtf==1.2.2
wtforms==3.2.1
email-validator==2.2.0
web3==7.6.1
sqlalchemy_utils==0.41.2